backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0
